fix(ci): Linux portability — ternary types, st_mtime, d_name
Three more GCC-only issues exposed by CI:

- `response.cpp`: `_stream->eof() ? RES_DONE : _state.raw()` mixes
  `enum ResponseState` (int) with the wrapper's `short`. GCC's
  `-Wextra` enforces `-Werror=enum-conversion` on the ternary. Both
  branches now cast to `int` explicitly.
- `status_codes.cpp`: `entry->d_namlen` is BSD/macOS-only — Linux's
  `dirent` doesn't have it. Replaced the empty-name check with
  `d_name[0] == '\0'`, which works on both.
- `status_codes.cpp`: `stat.st_mtimespec` is also Apple-only.
  `st_mtime` (a `time_t`) is POSIX and is enough for the directory
  listing's `strftime`.

+++ src/http/response.cpp
@@ -351,7 +351,7 @@ void Response::sendLengthedBody(const sockfd &fd) {
 
 	_stream->read(buff, (1 << 10));
 	::send(fd, buff, _stream->gcount(), 0);
-	setState(_stream->eof() ? RES_DONE : _state.raw());
+	setState(_stream->eof() ? (int)RES_DONE : (int)_state.raw());
 }
 
 void Response::setupChunkedBody() {
@@ -372,7 +372,7 @@ void Response::sendChunkedBody(const sockfd &fd) {
 		_headerBuffer.str("");
 		_headerBuffer.clear();
 	}
-	setState(_stream->eof() ? RES_DONE : _state.raw());
+	setState(_stream->eof() ? (int)RES_DONE : (int)_state.raw());
 }
 
 void Response::setupRangedBody() {
@@ -562,7 +562,7 @@ void Response::sendUploadBody(const sockfd &fd) {
 
 	_stream->read(buff, (1 << 10));
 	::send(fd, buff, _stream->gcount(), 0);
-	setState(_stream->eof() ? RES_DONE : _state.raw());
+	setState(_stream->eof() ? (int)RES_DONE : (int)_state.raw());
 }
 // =================================================================== Builder
 //
